feat: errorMessage 상태메시지 출력, 컴포넌트명 BookmarkRenderer -> StarredRenderer 로 변경 #83
